+++ includes/blocks/back-to-top/src/save.js
@@ -1,0 +1,9 @@
+import { useBlockProps } from '@wordpress/block-editor';
+
+export default function save() {
+    return (
+        <div {...useBlockProps.save()}>
+            <a href="#top" className="oes-back-to-top no-print" aria-label="Back to top"></a>
+        </div>
+    );
+}