+++ includes/blocks/print/src/index.js
@@ -3,6 +3,7 @@ import Edit from './edit';
 import save from './save';
 import metadata from './block.json';
 import oesIcon from './../../icon';
+import './style.css';
 
 registerBlockType(metadata.name, {
 	icon: oesIcon,