+++ includes/theme/filter/filter.js
@@ -227,7 +227,9 @@
     /** Initialize filter. */
     oesFilter.init = function () {
 
-        if (typeof oes_filter === 'undefined') return;
+        if (typeof oes_filter === 'undefined') {
+            oes_filter = [];
+        }
 
         const urlParams = new URLSearchParams(window.location.search);
         const params = Object.fromEntries(urlParams.entries());